--- src/main/java/com/vintly/common/config/SecurityUtil.java
@@ -1,25 +1,0 @@
-package com.vintly.common.config;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class SecurityUtil {
-
-    // 현재 로그인한 사용자 이름 가져오기
-    public static String getCurrentUsername() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if (authentication == null || !authentication.isAuthenticated()) {
-            return "";  // 인증되지 않은 사용자
-        }
-
-        Object principal = authentication.getPrincipal();
-
-        if (principal instanceof UserDetails) {
-            return ((UserDetails) principal).getUsername();
-        } else {
-            return principal.toString();
-        }
-    }
-}
