--- src/main/java/com/vintly/common/jwt/JWTFilter.java
@@ -1,66 +1,0 @@
-package com.vintly.common.jwt;
-
-import com.vintly.common.exception.memebr.AccessExpiredException;
-import com.vintly.entity.Member;
-import com.vintly.member.service.CustomUserDetails;
-import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-public class JWTFilter extends OncePerRequestFilter {
-
-    private final JWTUtil jwtUtil;
-
-    public JWTFilter(JWTUtil jwtUtil) {
-        this.jwtUtil = jwtUtil;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-
-        // 헤더에서 access 키에 담긴 토큰을 꺼냄
-        String accessToken = request.getHeader("access");
-
-        // 토큰이 없다면 다음 필터로 넘김
-        if (accessToken == null){
-            filterChain.doFilter(request, response);
-            return ;
-        }
-
-        // 토큰 만료 여부 확인, 만료시 401
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-           throw new AccessExpiredException();
-        }
-
-        // 토큰이 access 인지 확인 (발급시 페이로드에 명시)
-        String category = jwtUtil.getCategory(accessToken);
-
-        if (!category.equals("access")) throw new AccessExpiredException();
-
-        // username, role 값을 획득
-        String username = jwtUtil.getUsername(accessToken);
-        String role = jwtUtil.getRole(accessToken);
-
-        Member member = Member.builder()
-                .email(username)
-                .role(role)
-                .build();
-
-        CustomUserDetails customUserDetails = new CustomUserDetails(member);
-
-        Authentication authToken = new UsernamePasswordAuthenticationToken(customUserDetails, null, customUserDetails.getAuthorities());
-        SecurityContextHolder.getContext().setAuthentication(authToken);
-
-        filterChain.doFilter(request, response);
-    }
-}
