+++ src/main/java/com/vintly/domain/member/service/CustomUserDetailsService.java
@@ -1,0 +1,31 @@
+package com.vintly.domain.member.service;
+
+import com.vintly.domain.member.entity.Member;
+import com.vintly.domain.member.repo.MemberRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Autowired
+    public CustomUserDetailsService(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+
+        //System.out.println("username: " + username);
+        // 이메일을 통해 사용자 정보를 조회하고, 없으면 예외를 던짐 (401)
+        Member member = memberRepository.findByEmail(username)
+                .orElseThrow(() -> new UsernameNotFoundException(""));
+
+        return new CustomUserDetails(member);
+    }
+}
