+++ src/main/java/com/vintly/infra/config/redis/RedisCacheConfig.java
@@ -1,0 +1,33 @@
+package com.vintly.infra.config.redis;
+
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableCaching // Spring Boot의 캐싱 설정을 활성화
+public class RedisCacheConfig {
+
+    /*@Bean 강의에서는 controller 에서 사용
+    public CacheManager refreshCacheManager(RedisConnectionFactory redisConnectionFactory) {
+        RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration
+                .defaultCacheConfig()
+                // Redis에 Key를 저장할 때 String으로 직렬화(변환)해서 저장
+                .serializeKeysWith(
+                        RedisSerializationContext.SerializationPair.fromSerializer(
+                                new StringRedisSerializer()))
+                // Redis에 Value를 저장할 때 Json으로 직렬화(변환)해서 저장
+                .serializeValuesWith(
+                        RedisSerializationContext.SerializationPair.fromSerializer(
+                                new Jackson2JsonRedisSerializer<Object>(Object.class)
+                        )
+                )
+                // 데이터의 만료기간(TTL) 설정
+                .entryTtl(Duration.ofDays(1L));
+
+        return RedisCacheManager
+                .RedisCacheManagerBuilder
+                .fromConnectionFactory(redisConnectionFactory)
+                .cacheDefaults(redisCacheConfiguration)
+                .build();
+    }*/
+}
