+++ src/main/java/com/vintly/infra/member/MemberRepositoryImpl.java
@@ -1,0 +1,44 @@
+package com.vintly.infra.member;
+
+import com.vintly.domain.member.entity.Member;
+import com.vintly.domain.member.model.constant.Use;
+import com.vintly.domain.member.repo.MemberRepository;
+import org.springframework.stereotype.Repository;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+@Repository
+public class MemberRepositoryImpl implements MemberRepository {
+
+    private final MemberJpaRepository memberJpaRepository;
+
+    public MemberRepositoryImpl(MemberJpaRepository memberJpaRepository) {
+        this.memberJpaRepository = memberJpaRepository;
+    }
+
+    @Override
+    public Boolean existsByEmail(String email) {
+        return memberJpaRepository.existsByEmail(email);
+    }
+
+    @Override
+    public Boolean existsByNickname(String nickname) {
+        return memberJpaRepository.existsByNickname(nickname);
+    }
+
+    @Override
+    public Integer deleteByCreateDateBeforeAndUseYn(LocalDateTime today, Use use) {
+        return memberJpaRepository.deleteByCreateDateBeforeAndUseYn(today, use);
+    }
+
+    @Override
+    public Optional<Member> findByEmailCodeAndEmail(String code, String email) {
+        return memberJpaRepository.findByEmailCodeAndEmail(code, email);
+    }
+
+    @Override
+    public Optional<Member> findByEmail(String email) {
+        return memberJpaRepository.findByEmail(email);
+    }
+}
