+++ src/main/java/com/vintly/interfaces/filter/jwt/CustomLogoutFilter.java
@@ -1,0 +1,90 @@
+package com.vintly.interfaces.filter.jwt;
+
+import com.vintly.infra.config.jwt.JWTUtil;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.data.redis.core.StringRedisTemplate;
+import org.springframework.web.filter.GenericFilterBean;
+
+import java.io.IOException;
+
+public class CustomLogoutFilter extends GenericFilterBean {
+
+    private final JWTUtil jwtUtil;
+    private final StringRedisTemplate redisTemplate;
+
+    public CustomLogoutFilter(JWTUtil jwtUtil, StringRedisTemplate redisTemplate) {
+        this.jwtUtil = jwtUtil;
+        this.redisTemplate = redisTemplate;
+    }
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
+        doFilter((HttpServletRequest) request, (HttpServletResponse) response, chain);
+    }
+
+    private void doFilter(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws IOException, ServletException {
+
+        // 요청 경로 및 메서드 확인
+        if (!"/logout".equals(request.getRequestURI()) || !"POST".equalsIgnoreCase(request.getMethod())) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        // refresh 토큰 가져오기 (쿠키가 없을 경우 NPE 방지)
+        String refresh = null;
+        Cookie[] cookies = request.getCookies();
+        if(cookies != null){
+            for (Cookie cookie : cookies) {
+                if (cookie.getName().equals("refresh")) {
+                    refresh = cookie.getValue();
+                    break; // 더 이상 반복할 필요 없음
+                }
+            }
+        }
+
+        // refresh 토큰이 없으면 400 반환
+        if (refresh == null) {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            return;
+        }
+
+        // refresh 토큰 만료 여부 확인
+        try {
+            jwtUtil.isExpired(refresh);
+        } catch (ExpiredJwtException e) {
+            // 400
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            return;
+        }
+
+        // 토큰이 refresh 인지 확인 (발급시 페이로드에 명시)
+        String category = jwtUtil.getCategory(refresh);
+        if (!category.equals("refresh")) {
+            response.setStatus(HttpServletResponse.SC_BAD_REQUEST); // 400 code
+            return;
+        }
+
+        // [로그아웃] redis - refresh token 제거
+        String redisKey = "refresh:" + jwtUtil.getUsername(refresh);
+        //System.out.println(redisKey);
+        redisTemplate.delete(redisKey);
+
+        //Refresh 토큰 Cookie 값 0
+        Cookie cookie = new Cookie("refresh", null);
+        cookie.setMaxAge(0);
+        cookie.setPath("/");
+        cookie.setHttpOnly(true); // XSS 공격 방지, JS 접근 방지
+        // cookie.setSecure(true); // HTTPS에서만 전송
+
+        response.addCookie(cookie);
+        response.setStatus(HttpServletResponse.SC_OK); // 200
+    }
+
+}
