+++ src/main/java/com/vintly/interfaces/presentation/ApiResponse.java
@@ -1,0 +1,19 @@
+package com.vintly.interfaces.presentation;
+
+import com.vintly.interfaces.member.exception.StatusEnum;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class ApiResponse<T> {
+    private StatusEnum status;
+    private String msg;
+    private T data;
+
+    @Builder
+    public ApiResponse(StatusEnum status, String msg, T data){
+        this.status = status;
+        this.msg = msg;
+        this.data = data;
+    }
+}
