--- src/main/java/com/vintly/member/controller/MemberController.java
@@ -1,32 +1,0 @@
-package com.vintly.member.controller;
-
-import com.vintly.common.config.ApiResponse;
-import com.vintly.common.exception.StatusEnum;
-import com.vintly.common.exception.memebr.NicknameValidException;
-import com.vintly.member.model.req.JoinReq;
-import com.vintly.member.service.MemberService;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-
-import java.io.IOException;
-
-@RestController
-@RequestMapping("/api/v1/members")
-@Validated
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @Autowired
-    public MemberController(MemberService memberService){
-        this.memberService = memberService;
-    }
-
-}
